Download albums

+++ wp-content/plugins/member-register-type/controllers/MrtApiPhotos.php
@@ -60,7 +60,7 @@ class MrtApiPhotos extends WP_REST_Controller {
         register_rest_route($this->route->namespace, $this->route->base($this->rest_base, 'download_items'), array(
             'methods' => WP_REST_Server::CREATABLE,
             'callback' => array($this, 'download_items'),
-            // 'permission_callback' => array($this, 'items_permissions_check'),
+                // 'permission_callback' => array($this, 'items_permissions_check'),
         ));
     }
 
@@ -141,8 +141,10 @@ class MrtApiPhotos extends WP_REST_Controller {
             $filestack_validate = $this->validate_filestack($validate['input']);
 
             if ($filestack_validate['status']) {
-                $this->mrt_file_stack->delete('pf_photo_id', $this->mrt_photo->id);
-                $this->mrt_file_stack->insert($filestack_validate['input']);
+                if (isset($filestack_validate['input']['change_photo']) && $filestack_validate['input']['change_photo'] == true) {
+                    $this->mrt_file_stack->delete('pf_photo_id', $this->mrt_photo->id);
+                    $this->mrt_file_stack->insert($filestack_validate['input']);
+                }
             } else {
                 return new WP_REST_Response(['message' => $filestack_validate['message']], 400);
             }
@@ -161,7 +163,7 @@ class MrtApiPhotos extends WP_REST_Controller {
 
         if (isset($this->mrt_photo->id)) {
             $this->mrt_photo->delete();
-            
+
             $this->mrt_file_stack->update(['deleteFlag' => 1], 'pf_photo_id', $this->mrt_photo->id);
             return new WP_REST_Response([], 200);
         }
@@ -233,15 +235,15 @@ class MrtApiPhotos extends WP_REST_Controller {
     }
 
     public function download_items($request) {
-        
+
         $input = $request->get_params();
-        
+
         $photos = $this->mrt_photo->getset_items($input['ids']);
         $files = [];
         foreach ($photos as $key => $value) {
             $files[] = $value['original'];
         }
-        
+
         $zip['zip_url'] = State::create_zip($files);
         $zip['files_count'] = count($files);
         return new WP_REST_Response($zip, 200);
